+++ app/components/recording_studio_billing/plan_card_component.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module RecordingStudioBilling
+  class PlanCardComponent < BaseComponent
+    def initialize(plan:, presenter:)
+      super()
+      @plan = plan
+      @presenter = presenter
+    end
+  end
+end